Create html-table-to-csv
This does not produce "Official CSV", the default separator is actually tabs, when it should be comma's.  Easy to modify the parser object if needed.  Instead of storing the string of data inside it is immediately passed out to insure low memory consumption and greater, speed so it can be written to the display buffer immediately.  data is `escape`d on input and on output to ensure data safety. (not a HTML security expert.)

TODO:
 - Implement `<caption>` and `<colgroup>`
 - Add minimum `<thead>` and `<tfoot>` features e.g. horizontal divider line.
 - Manage multiple tables by `id`